+++ src/NFugue/Integration/MusicXml/Internals/PartContext.cs
@@ -5,8 +5,8 @@
         public string Id { get; set; }
         public string Name { get; set; }
         public MidiInstrument[] Instruments { get; set; } = new MidiInstrument[16];
-        public byte CurrentVolume { get; set; } = (byte)DefaultNoteSettings.DefaultOnVelocity;
-        public byte Voice { get; set; }
+        public int CurrentVolume { get; set; } = DefaultNoteSettings.DefaultOnVelocity;
+        public int Voice { get; set; }
 
         public PartContext(string id, string name)
         {
