+++ src/NFugue/Midi/MidiTools.cs
@@ -2,7 +2,7 @@
 {
     public class MidiTools
     {
-        public static sbyte GetLSB(int value) => (sbyte)(value & 0x7F);
-        public static sbyte GetMSB(int value) => (sbyte)((value >> 7) & 0x7F);
+        public static int GetLSB(int value) => (value & 0x7F);
+        public static int GetMSB(int value) => ((value >> 7) & 0x7F);
     }
 }
